drone

+++ dotNet5782_1974_0044/BL/BO/DroneInCharging.cs
@@ -11,7 +11,7 @@ namespace IBL
         public class DroneInCharging
         {
             public int Id { get; set; }
-            public int ChargingMode { get; set; }
+            public double ChargingMode { get; set; }
         }
     }
